# Implement raw tag parsing in BlockBody