--- a/ext/liquid_c/raw.h
+++ b/ext/liquid_c/raw.h
@@ -1,6 +1,0 @@
-#ifndef LIQUID_RAW_H
-#define LIQUID_RAW_H
-
-void liquid_define_raw();
-
-#endif