--- a/ext/liquid_c/tokenizer.h
+++ b/ext/liquid_c/tokenizer.h
@@ -26,6 +26,7 @@ typedef struct tokenizer {
     unsigned int line_number;
     bool lstrip_flag;
     bool for_liquid_tag;
+    bool for_raw_tag;
 
     // Temporary to test rollout of the fix for this bug
     bool bug_compatible_whitespace_trimming;