… ext/liquid_c/raw.c